[release] ログイン時にサーバーに存在するエネミー情報を取得する処理の実装,他バグ修正

+++ DxGameClient/program/game/Actor/Actor.cpp
@@ -326,7 +326,7 @@ void Actor::DefaultAttack()
 		//ƒ_ƒ[ƒW”»’è
 		bool successAttack = battleLogic->IsSuccessAttack(attackData, defendData, static_cast<int>(myType));
 
-		float damage = battleLogic->CalcDefaultDamage(attackData->GetAttack(), defendData->GetDefence(), attackData->GetLevel(), successAttack);
+		double damage = battleLogic->CalcDefaultDamage(attackData->GetAttack(), defendData->GetDefence(), attackData->GetLevel(), successAttack);
 		defendData->UpdateHp((damage * (-1)));
 	}
